general-invoice-parser/PdfParser/requirements.txt
attrs==21.4.0
boto3==1.22.9
botocore==1.25.9
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==37.0.2
idna==3.3
isodate==0.6.1
jmespath==1.0.0
lxml==4.8.0
netsuitesdk==2.12.0
numpy==1.22.3
pandas==1.4.2
pdfminer.six==20220319
pdfplumber==0.6.2
Pillow==9.1.0
platformdirs==2.5.2
pycparser==2.21
PyPDF2==1.27.12
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
Wand==0.6.7
zeep==4.1.0
